Added contact section to README
A new "Contact" section has been added to the README file. This includes information on how to reach out for questions, suggestions, or feedback. It also provides a link to join our Discord server and acknowledges the user's choice of our Help Desk service. A closing note indicating the creator of the service is also included.